feat: Account 엔티티 필드 자료형 변경

+++ src/main/java/com/naver/line/demo/account/entities/Account.java
@@ -19,7 +19,7 @@ import java.time.LocalDateTime;
 public class Account {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column
     private Integer user_id;
@@ -28,7 +28,7 @@ public class Account {
     private String number;
 
     @Column
-    private Integer amount; // TODO - check boundary
+    private Long amount;
 
     @Column
     private User.Status status;
